refactor(operations): extract useDayFeed hook from DailyLog — entity_id + raw actuals on task rows, scene durations selected
