+++ src/components/workbench/operations/content/DailyLog.tsx
@@ -9,84 +9,35 @@
  *     READ-ONLY here; the proof is committed on the Daily Plan tab (daily_plan_items
  *     → calendar_blocks). This is the execution record CE-5 will script from.
  *
- * Ordering: timed rows (blocks: actual??scheduled; scenes: step time_of_day) sort by
- * minute-of-day; untimed scenes fall to the end by step_order (no fabricated times).
+ * The read + interleave (fetches, merge, DAY_START order) now live in the shared
+ * useDayFeed(date) hook (so the day calendar consumes the SAME feed and can't drift).
+ * DailyLog owns only the answer-write UI (per-scene drafts, save, start-day) — those
+ * mutate the hook's cells/pieces via the exposed setters. Zero behavior change.
  *
- * Zero new write paths: reads /content/grid (scenes/pieces/cells),
- * /daily-plan/items?from=&to= (committed blocks), /operations/projects (names);
- * writes only via the existing /content/grid/cell (answer) + /content/grid/piece
- * (start day). No AI (CE-5). Storage unchanged: the answer is take.script.
+ * Zero new write paths: writes only via the existing /content/grid/cell (answer) +
+ * /content/grid/piece (start day). No AI (CE-5). Storage unchanged: answer = take.script.
  */
 
 'use client';
 
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useState } from 'react';
 import { useOperationsEntity } from '../EntitySelector';
-import { CONTENT_DAY_PLAN_CHANGED_EVENT } from './ScenifyModal';
 import TaskBand from './TaskBand';
 import {
-  compareDayOrder,
-  minuteOfDayFromInstant,
-  minuteOfDayFromTime,
-} from '@/lib/content/dayOrder';
-
-interface RoutineStepLite {
-  id: string;
-  step_order: number;
-  activity: string;
-  time_of_day: string | null;
-  routine_id: string;
-}
-interface SceneRow {
-  id: string;
-  entity_id: string;
-  assigned_question_text: string | null;
-  narrative_purpose: string | null;
-  b_roll: string | null;
-  routine_step: RoutineStepLite;
-}
-interface PieceCol {
-  id: string;
-  entity_id: string;
-  piece_date: string;
-}
-interface Cell {
-  id: string;
-  scene_id: string;
-  piece_id: string;
-  script: string | null;
-}
-interface CalendarBlock {
-  id: string;
-  scheduled_start: string;
-  scheduled_end: string;
-  actual_start: string | null;
-  actual_end: string | null;
-  status: string;
-}
-interface PlanItem {
-  id: string;
-  entity_id: string;
-  ad_hoc_title: string | null;
-  task: { id: string; title: string; project_id: string; status: string } | null;
-  calendar_blocks: CalendarBlock[];
-}
+  useDayFeed,
+  type SceneRow,
+  type Cell,
+  type DayTaskBlock,
+} from './useDayFeed';
 
 type RowState = 'idle' | 'saving' | 'saved' | 'error';
 
-const dayOf = (iso: string) => iso.slice(0, 10);
-
 // "1970-01-01T07:30:00Z" (Prisma @db.Time) → "07:30" for DISPLAY. Sort minutes come
-// from the shared dayOrder helper so S3 and the record grid can never drift.
+// from the shared dayOrder helper (in useDayFeed) so S3 and the record grid can't drift.
 const fmtTimeOfDay = (t: string | null): string => {
   if (!t) return '';
   const m = t.match(/T(\d{2}:\d{2})/);
   return m ? m[1] : '';
-};
-// Timestamptz instant → local wall-clock "HH:MM" for DISPLAY.
-const fmtClock = (iso: string): string => {
-  const d = new Date(iso);
-  return `${String(d.getHours()).padStart(2, '0')}:${String(d.getMinutes()).padStart(2, '0')}`;
 };
 
 const headerCellClass =
@@ -96,201 +47,28 @@ export default function DailyLog({ date }: { date: string }) {
   // selectedEntityId is used ONLY to create the day's canonical piece. Reading is
   // CROSS-ENTITY (OPS-CE-8): the day is ONE reel — scenes/answers/blocks span entities.
   const { selectedEntityId } = useOperationsEntity();
-  const [scenes, setScenes] = useState<SceneRow[] | null>(null);
-  const [pieces, setPieces] = useState<PieceCol[] | null>(null);
-  const [cells, setCells] = useState<Cell[]>([]);
-  const [projectNameById, setProjectNameById] = useState<Record<string, string>>({});
-  const [planItems, setPlanItems] = useState<PlanItem[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+
+  // The shared day feed: cross-entity scenes/answers + the day's committed blocks,
+  // merged + DAY_START-ordered. setCells/setPieces/setError back the write UI below.
+  const {
+    timeline,
+    dayScenes,
+    piece,
+    cellByScene,
+    answeredCount,
+    loading,
+    error,
+    setError,
+    setCells,
+    setPieces,
+  } = useDayFeed(date);
+
   const [starting, setStarting] = useState(false);
 
   // Per-row local answer drafts + save state (scene rows only).
   const [drafts, setDrafts] = useState<Record<string, string>>({});
   const [rowState, setRowState] = useState<Record<string, RowState>>({});
   const [rowError, setRowError] = useState<Record<string, string>>({});
-
-  // Grid (scenes/pieces/cells) + project names — CROSS-ENTITY (no entity filter):
-  // the day's reel reads every entity's scenes/answers.
-  const load = useCallback(async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const [gridRes, projRes] = await Promise.all([
-        fetch('/api/operations/content/grid', { credentials: 'include' }),
-        fetch('/api/operations/projects', { credentials: 'include' }),
-      ]);
-      if (!gridRes.ok) throw new Error(`Failed to load (${gridRes.status})`);
-      const body = await gridRes.json();
-      setScenes(body.scenes ?? []);
-      setPieces(body.pieces ?? []);
-      setCells(body.cells ?? []);
-      if (projRes.ok) {
-        const pb = await projRes.json();
-        const map: Record<string, string> = {};
-        for (const p of pb.projects ?? []) map[p.id] = p.title;
-        setProjectNameById(map);
-      }
-    } catch (e) {
-      setError(e instanceof Error ? e.message : 'failed to load daily log');
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  // Committed task blocks for the selected day (read-only execution record).
-  const loadBlocks = useCallback(async () => {
-    try {
-      const res = await fetch(
-        `/api/operations/daily-plan/items?from=${date}&to=${date}`,
-        { credentials: 'include' }
-      );
-      if (!res.ok) return;
-      const body = await res.json();
-      setPlanItems(body.items ?? []);
-    } catch {
-      /* leave prior blocks on a transient failure */
-    }
-  }, [date]);
-
-  useEffect(() => {
-    void load();
-  }, [load]);
-  useEffect(() => {
-    void loadBlocks();
-  }, [loadBlocks]);
-  // OPS-CE-8D: re-read the day's tasks when one is added on the day (S1 add-to-day).
-  useEffect(() => {
-    const refresh = () => void loadBlocks();
-    window.addEventListener(CONTENT_DAY_PLAN_CHANGED_EVENT, refresh);
-    return () => window.removeEventListener(CONTENT_DAY_PLAN_CHANGED_EVENT, refresh);
-  }, [loadBlocks]);
-
-  // CROSS-ENTITY: every active scene for the day, regardless of entity.
-  const dayScenes = useMemo(
-    () =>
-      (scenes ?? [])
-        .slice()
-        .sort((a, b) => a.routine_step.step_order - b.routine_step.step_order),
-    [scenes]
-  );
-
-  // The day's CANONICAL piece: ONE piece per date = ONE reel. Resolve the first
-  // piece for the date across ALL entities (grid GET orders by piece_date, created_at),
-  // so cross-entity answers attach to a single day column. (Cells allow scene.entity ≠
-  // piece.entity — the cell route checks user ownership, not entity match.)
-  const piece = useMemo(
-    () => (pieces ?? []).find((p) => dayOf(p.piece_date) === date) ?? null,
-    [pieces, date]
-  );
-
-  const cellByScene = useMemo(() => {
-    const m = new Map<string, Cell>();
-    if (piece) for (const c of cells) if (c.piece_id === piece.id) m.set(c.scene_id, c);
-    return m;
-  }, [cells, piece]);
-
-  const answeredCount = useMemo(
-    () => dayScenes.filter((s) => (cellByScene.get(s.id)?.script ?? '').trim().length > 0).length,
-    [dayScenes, cellByScene]
-  );
-
-  // Flatten plan items into read-only task rows: one TIMED row per calendar block,
-  // plus — OPS-CE-8C — one UNTIMED "planned" row for any item with NO block yet
-  // (added via "+ add to day" but not time-committed). Untimed rows sink after the
-  // untimed scenes (large order base) and order among themselves by plan order, via
-  // the CE-8B shared comparator.
-  const UNTIMED_TASK_ORDER_BASE = 100000;
-  const taskBlocks = useMemo(() => {
-    const rows: {
-      id: string;
-      itemId: string;
-      blockId: string | null;
-      taskId: string | null;
-      projectId: string | null;
-      title: string;
-      projectName: string | null;
-      status: string;
-      scheduledStart: string | null;
-      scheduledEnd: string | null;
-      label: string;
-      minute: number | null;
-      order: number;
-      planned: boolean;
-    }[] = [];
-    let plannedIndex = 0;
-    for (const item of planItems) {
-      // CROSS-ENTITY: all items for the date (personal + business).
-      const title = item.task?.title ?? item.ad_hoc_title ?? 'Untitled';
-      const projectName = item.task?.project_id ? projectNameById[item.task.project_id] ?? null : null;
-      const taskId = item.task?.id ?? null;
-      const projectId = item.task?.project_id ?? null;
-      if (item.calendar_blocks.length === 0) {
-        // Assigned to the day but no time committed yet — visible immediately.
-        rows.push({
-          id: `item-${item.id}`,
-          itemId: item.id,
-          blockId: null,
-          taskId,
-          projectId,
-          title,
-          projectName,
-          status: 'planned',
-          scheduledStart: null,
-          scheduledEnd: null,
-          label: '',
-          minute: null,
-          order: UNTIMED_TASK_ORDER_BASE + plannedIndex++,
-          planned: true,
-        });
-        continue;
-      }
-      for (const b of item.calendar_blocks) {
-        const useActual = !!b.actual_start;
-        const start = useActual ? (b.actual_start as string) : b.scheduled_start;
-        const end = useActual ? b.actual_end : b.scheduled_end;
-        const label = `${fmtClock(start)}–${end ? fmtClock(end) : '…'} ${useActual ? '(actual)' : '(scheduled)'}`;
-        const minute = minuteOfDayFromInstant(start);
-        rows.push({
-          id: b.id,
-          itemId: item.id,
-          blockId: b.id,
-          taskId,
-          projectId,
-          title,
-          projectName,
-          status: b.status,
-          scheduledStart: b.scheduled_start,
-          scheduledEnd: b.scheduled_end,
-          label,
-          minute,
-          order: minute,
-          planned: false,
-        });
-      }
-    }
-    return rows;
-  }, [planItems, projectNameById]);
-
-  // Merge-sort scenes + task blocks into one timeline. Timed rows by minute-of-day;
-  // untimed scenes (no step time) sink to the end by step_order (no fabricated time).
-  type TimelineRow =
-    | { kind: 'scene'; minute: number | null; order: number; scene: SceneRow }
-    | { kind: 'task'; minute: number | null; order: number; block: (typeof taskBlocks)[number] };
-  const timeline = useMemo<TimelineRow[]>(() => {
-    const rows: TimelineRow[] = [
-      ...dayScenes.map((s) => ({
-        kind: 'scene' as const,
-        minute: minuteOfDayFromTime(s.routine_step.time_of_day),
-        order: s.routine_step.step_order,
-        scene: s,
-      })),
-      ...taskBlocks.map((b) => ({ kind: 'task' as const, minute: b.minute, order: b.order, block: b })),
-    ];
-    // The ONE shared day-anchored order (midnight wraps to day-end).
-    rows.sort(compareDayOrder);
-    return rows;
-  }, [dayScenes, taskBlocks]);
 
   const draftFor = (sceneId: string) =>
     drafts[sceneId] !== undefined ? drafts[sceneId] : cellByScene.get(sceneId)?.script ?? '';
@@ -426,7 +204,7 @@ export default function DailyLog({ date }: { date: string }) {
   };
 
   // Read-only execution band — fully legible, labeled, wrapped (shared TaskBand).
-  const renderTaskRow = (b: (typeof taskBlocks)[number]) => (
+  const renderTaskRow = (b: DayTaskBlock) => (
     <tr key={`task-${b.id}`}>
       <td colSpan={6} className="border border-border-light border-l-4 border-l-amber-400 bg-amber-50/50 px-3 py-2 align-top">
         <TaskBand
@@ -508,7 +286,7 @@ export default function DailyLog({ date }: { date: string }) {
               {timeline.map((row) =>
                 row.kind === 'scene' ? renderSceneRow(row.scene) : renderTaskRow(row.block)
               )}
-              {taskBlocks.length === 0 && (
+              {!timeline.some((row) => row.kind === 'task') && (
                 <tr>
                   <td colSpan={6} className="border border-border-light px-3 py-1.5 text-text-muted">
                     no task blocks committed — assign tasks in section 1 or on the Daily Plan tab
